Rename package from `aliaroaccessoryboards` to `accessoryboards`
Updated all imports, paths, and references in the codebase and tests to reflect the new package name. Adjusted project files, including `pyproject.toml` and `README.md`, to ensure consistency.
