+++ accessoryboards/boards/32ch_instrumentation_switch.brd
