+++ accessoryboards/examples/exceptions/exclusive_connection_conflict.py
@@ -1,8 +1,8 @@
-from aliaroaccessoryboards import (
-    PathUnsupportedException,
-    AccessoryBoard,
+from accessoryboards import (
     BoardConfig,
+    AccessoryBoard,
     SimulatedBoardController,
+    ExclusiveConnectionConflictException,
 )
 
 # Step 1: Create a configuration for the board
@@ -18,6 +18,7 @@ board_config = BoardConfig.from_device_name("32ch_instrumentation_switch")
 board = AccessoryBoard(board_config, SimulatedBoardController(board_config))
 
 try:
-    board.connect_channels("DUT_CH01", "DUT_CH02")
-except PathUnsupportedException as e:
+    board.connect_channels("DUT_CH01", "BUS_POS")
+    board.connect_channels("DUT_CH01", "BUS_NEG")
+except ExclusiveConnectionConflictException as e:
     print(e)
