+++ accessoryboards/examples/exceptions/source_conflict.py
@@ -1,8 +1,8 @@
-from aliaroaccessoryboards import (
+from accessoryboards import (
     BoardConfig,
     AccessoryBoard,
     SimulatedBoardController,
-    ExclusiveConnectionConflictException,
+    SourceConflictException,
 )
 
 # Step 1: Create a configuration for the board
@@ -18,7 +18,9 @@ board_config = BoardConfig.from_device_name("32ch_instrumentation_switch")
 board = AccessoryBoard(board_config, SimulatedBoardController(board_config))
 
 try:
+    board.mark_as_source("DUT_CH01")
+    board.mark_as_source("DUT_CH02")
     board.connect_channels("DUT_CH01", "BUS_POS")
-    board.connect_channels("DUT_CH01", "BUS_NEG")
-except ExclusiveConnectionConflictException as e:
+    board.connect_channels("DUT_CH02", "BUS_POS")
+except SourceConflictException as e:
     print(e)
